refactor: DiceSystemPicker を共通コンポーネント化、SettingsModal にも適用
